Windows support for sdist build. Will fix the OSX build errors on master.

+++ cpp/cpp_odbc/Library/cpp_odbc/connection.h
@@ -35,11 +35,11 @@ public:
 	std::shared_ptr<statement const> make_statement() const;
 
 	/**
-	 * @brief Set the attribute to the given long value
+	 * @brief Set the attribute to the given intptr_t value
 	 * @param attribute An ODBC constant which represents the attribute which shall be set
 	 * @param value The new value for the attribute
 	 */
-	void set_attribute(SQLINTEGER attribute, long value) const;
+	void set_attribute(SQLINTEGER attribute, intptr_t value) const;
 
 	/**
 	 * @brief End the current transaction by committing all changes to the database
@@ -77,7 +77,7 @@ protected:
 	connection();
 private:
 	virtual std::shared_ptr<statement const> do_make_statement() const = 0;
-	virtual void do_set_attribute(SQLINTEGER attribute, long value) const = 0;
+	virtual void do_set_attribute(SQLINTEGER attribute, intptr_t value) const = 0;
 	virtual void do_commit() const = 0;
 	virtual void do_rollback() const = 0;
 	virtual std::string do_get_string_info(SQLUSMALLINT info_type) const = 0;
