+++ cpp/cpp_odbc/Library/cpp_odbc/level2/level1_connector.h
@@ -36,8 +36,8 @@ private:
 	diagnostic_record do_get_diagnostic_record(statement_handle const & handle) const final;
 	diagnostic_record do_get_diagnostic_record(connection_handle const & handle) const final;
 	diagnostic_record do_get_diagnostic_record(environment_handle const & handle) const final;
-	void do_set_environment_attribute(environment_handle const & handle, SQLINTEGER attribute, long value) const final;
-	void do_set_connection_attribute(connection_handle const & handle, SQLINTEGER attribute, long value) const final;
+	void do_set_environment_attribute(environment_handle const & handle, SQLINTEGER attribute, intptr_t value) const final;
+	void do_set_connection_attribute(connection_handle const & handle, SQLINTEGER attribute, intptr_t value) const final;
 	void do_establish_connection(connection_handle & handle, std::string const & connection_string) const final;
 	void do_disconnect(connection_handle & handle) const final;
 	void do_end_transaction(connection_handle const & handle, SQLSMALLINT completion_type) const final;
@@ -49,14 +49,14 @@ private:
 	void do_execute_statement(statement_handle const & handle, std::string const & sql) const final;
 	bool do_fetch_scroll(statement_handle const & statement_handle, SQLSMALLINT fetch_orientation, SQLLEN fetch_offset) const final;
 	void do_free_statement(statement_handle const & handle, SQLUSMALLINT option) const final;
-	long do_get_integer_column_attribute(statement_handle const & handle, SQLUSMALLINT column_id, SQLUSMALLINT field_identifier) const final;
-	long do_get_integer_statement_attribute(statement_handle const & handle, SQLINTEGER attribute) const final;
+	intptr_t do_get_integer_column_attribute(statement_handle const & handle, SQLUSMALLINT column_id, SQLUSMALLINT field_identifier) const final;
+	intptr_t do_get_integer_statement_attribute(statement_handle const & handle, SQLINTEGER attribute) const final;
 	std::string do_get_string_column_attribute(statement_handle const & handle, SQLUSMALLINT column_id, SQLUSMALLINT field_identifier) const final;
 	short int do_number_of_result_columns(statement_handle const & handle) const final;
 	short int do_number_of_parameters(statement_handle const & handle) const final;
 	void do_prepare_statement(statement_handle const & handle, std::string const & sql) const final;
 	void do_prepare_statement(statement_handle const & handle, std::u16string const & sql) const final;
-	void do_set_statement_attribute(statement_handle const & handle, SQLINTEGER attribute, long value) const final;
+	void do_set_statement_attribute(statement_handle const & handle, SQLINTEGER attribute, intptr_t value) const final;
 	void do_set_statement_attribute(statement_handle const & handle, SQLINTEGER attribute, SQLULEN * pointer) const final;
 	SQLLEN do_row_count(statement_handle const & handle) const final;
 	column_description do_describe_column(statement_handle const & handle, SQLUSMALLINT column_id) const final;
