+++ cpp/cpp_odbc/Library/cpp_odbc/level3/raii_environment.h
@@ -43,7 +43,7 @@ public:
 
 private:
 	std::shared_ptr<connection> do_make_connection(std::string const & connection_string) const final;
-	void do_set_attribute(SQLINTEGER attribute, long value) const final;
+	void do_set_attribute(SQLINTEGER attribute, intptr_t value) const final;
 
 	struct intern;
 	std::unique_ptr<intern> impl_;
