+++ cpp/cpp_odbc/Library/src/connection.cpp
@@ -10,7 +10,7 @@ std::shared_ptr<statement const> connection::make_statement() const
 	return do_make_statement();
 }
 
-void connection::set_attribute(SQLINTEGER attribute, long value) const
+void connection::set_attribute(SQLINTEGER attribute, intptr_t value) const
 {
 	do_set_attribute(attribute, value);
 }
