+++ cpp/cpp_odbc/Library/src/level3/raii_environment.cpp
@@ -52,7 +52,7 @@ std::shared_ptr<connection> raii_environment::do_make_connection(std::string con
 	return std::make_shared<raii_connection>(as_raii_environment, connection_string);
 }
 
-void raii_environment::do_set_attribute(SQLINTEGER attribute, long value) const
+void raii_environment::do_set_attribute(SQLINTEGER attribute, intptr_t value) const
 {
 	impl_->api->set_environment_attribute(impl_->handle, attribute, value);
 }
