+++ cpp/cpp_odbc/Test/cpp_odbc_test/mock_connection.h
@@ -20,7 +20,7 @@ namespace cpp_odbc_test {
 	class mock_connection : public cpp_odbc::connection {
 	public:
 		MOCK_CONST_METHOD0( do_make_statement, std::shared_ptr<cpp_odbc::statement const>());
-		MOCK_CONST_METHOD2( do_set_attribute, void(SQLINTEGER, long));
+		MOCK_CONST_METHOD2( do_set_attribute, void(SQLINTEGER, intptr_t));
 		MOCK_CONST_METHOD0( do_commit, void());
 		MOCK_CONST_METHOD0( do_rollback, void());
 		MOCK_CONST_METHOD1( do_get_string_info, std::string(SQLUSMALLINT info_type));
